MPT-16437 Update http/mixins file structure
